Base du Unit Pulling
changement dans l'archi du pulling par rapport à la précédente commit,
Spawn manager se charge plus que du pulling et c'est LevelManager qui se
charge d'init les unit à spawn toutes les x sec selon le bpm ( constante
pour le moment) et leur charge des stats selon l'intensité.

+++ Unity/Dreamcatcher/Assets/Script/Manager/UnitStatManager.cs
@@ -14,9 +14,9 @@ public class UnitStatManager : MonoBehaviour {
 
 	public void InitStats ( EntityStatisticScript script, float levelIntensity)
 	{
-		script.Damage = levelIntensity * 4f * StatVariation() ;
-		script.CurrentHealth = levelIntensity * 20f  * StatVariation();
-		script.Resistance = levelIntensity * 2f  * StatVariation();
+		script.Damage = (int)(levelIntensity * 4f * StatVariation()) ;
+		script.CurrentHealth = (int) (levelIntensity * 20f  * StatVariation() );
+		script.Resistance = (int) (levelIntensity * 2f  * StatVariation() );
 	}
 
 	private float StatVariation()
